- Split off the entire translation code from r_draw.cpp into r_translate.cpp. - Moved the common code of ACS and DECORATE translation generation into the   FRemapTable class. - Fixed: The DECORATE translation code was not changed for the new data structures. - Expanded range of ACS and DECORATE translations to 65535. - Fixed: R_CopyTranslation was not altered for the new functionality.   I removed the function and replaced the one use with a simple assignment.
SVN r644 (trunk)

+++ src/actor.h
@@ -714,7 +714,7 @@ public:
 
 	BYTE FloatBobPhase;
 	BYTE FriendPlayer;				// [RH] Player # + 1 this friendly monster works for (so 0 is no player, 1 is player 0, etc)
-	WORD Translation;
+	DWORD Translation;
 
 	// [RH] Stuff that used to be part of an Actor Info
 	WORD SeeSound;
